# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Woxsom Code API - Multi-agent AI coding assistant
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiKeyInput {
  /**
     * Array of 1-5 Groq API keys
     * @minItems 1
     * @maxItems 5
     */
  keys: string[];
}

export interface ApiKeyStatus {
  configured: boolean;
  keyCount: number;
  maskedKeys?: string[];
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  idle: 'idle',
  planning: 'planning',
  executing: 'executing',
  reviewing: 'reviewing',
  done: 'done',
  error: 'error',
} as const;

export interface Session {
  id: string;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
  status: SessionStatus;
  hasProject?: boolean;
}

export interface SessionInput {
  title: string;
}

export type SessionDetailStatus = typeof SessionDetailStatus[keyof typeof SessionDetailStatus];


export const SessionDetailStatus = {
  idle: 'idle',
  planning: 'planning',
  executing: 'executing',
  reviewing: 'reviewing',
  done: 'done',
  error: 'error',
} as const;

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
  planner: 'planner',
  executor_frontend: 'executor_frontend',
  executor_backend: 'executor_backend',
  critic: 'critic',
} as const;

export interface Message {
  id: string;
  sessionId: string;
  role: MessageRole;
  content: string;
  createdAt: string;
  /** @nullable */
  agentName?: string | null;
  /** @nullable */
  metadata?: string | null;
}

export interface SessionDetail {
  id: string;
  title: string;
  createdAt: string;
  updatedAt: string;
  status: SessionDetailStatus;
  messages: Message[];
  hasProject?: boolean;
  projectFiles?: string[];
}

export interface MessageInput {
  content: string;
}

export type PipelineStatusStatus = typeof PipelineStatusStatus[keyof typeof PipelineStatusStatus];


export const PipelineStatusStatus = {
  idle: 'idle',
  planning: 'planning',
  executing: 'executing',
  reviewing: 'reviewing',
  done: 'done',
  error: 'error',
} as const;

export type AgentStatusGroup = typeof AgentStatusGroup[keyof typeof AgentStatusGroup];


export const AgentStatusGroup = {
  A: 'A',
  B: 'B',
  C: 'C',
} as const;

export type AgentStatusStatus = typeof AgentStatusStatus[keyof typeof AgentStatusStatus];


export const AgentStatusStatus = {
  idle: 'idle',
  active: 'active',
  done: 'done',
  error: 'error',
} as const;

export interface AgentStatus {
  name: string;
  group: AgentStatusGroup;
  status: AgentStatusStatus;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  task?: string | null;
}

export interface PipelineStatus {
  sessionId: string;
  status: PipelineStatusStatus;
  /** @nullable */
  currentStep?: string | null;
  agents: AgentStatus[];
  /** 0-100 progress percentage */
  progress?: number;
}

export type ModelInfoGroup = typeof ModelInfoGroup[keyof typeof ModelInfoGroup];


export const ModelInfoGroup = {
  reasoning: 'reasoning',
  fast: 'fast',
  specialized: 'specialized',
} as const;

export interface ModelInfo {
  id: string;
  name: string;
  group: ModelInfoGroup;
  capability: string;
  recommended?: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}