# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Woxsom Code API - Multi-agent AI coding assistant
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{sessionId}/download:
    get:
      operationId: downloadProject
      tags: [sessions]
      summary: Download generated project as ZIP
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ZIP file
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "404":
          description: No project generated yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...